app/javascript/lca/components/fields/TextField.test.tsx
import { fireEvent, render } from '@testing-library/react'

import LcaTextField from './TextField'

describe('LcaTextField', () => {
  beforeEach(() => {
    vi.useFakeTimers()
  })
  afterEach(() => {
    vi.restoreAllMocks()
  })

  test('renders without errors', () => {
    render(<LcaTextField onChange={() => null} />)
    // Add assertions here to verify that the component renders correctly
  })

  test('calls onChange handler when input value changes', () => {
    const handleChange = vi.fn()
    const { getByRole } = render(<LcaTextField onChange={handleChange} />)
    const input = getByRole('textbox')

    fireEvent.change(input, { target: { value: 'Test' } })
    vi.advanceTimersToNextTimer()

    expect(handleChange).toHaveBeenCalledTimes(1)
    expect(handleChange).toHaveBeenCalledWith(expect.any(Object))
    // Add more assertions here to verify the behavior of the onChange handler
  })

  // Add more tests here to cover other scenarios and edge cases
})
